refactor(cli): break down monolithic CLI into separate modules
